+++ apps/web/src/components/access-data/BAccessData.vue
@@ -1,22 +1,16 @@
 <script setup lang="ts">
-import { InformationCircle, Pause, Play } from "@vicons/ionicons5";
-import { Icon } from "@vicons/utils";
-import { useWebSocket, watchDebounced } from "@vueuse/core";
 import {
-	NButton,
-	NDataTable,
-	NForm,
-	NInput,
-	NInputNumber,
-	NPagination,
-	NSelect,
-	NTag,
-	NTooltip,
-	useNotification,
-} from "naive-ui";
+	breakpointsTailwind,
+	useBreakpoints,
+	useWebSocket,
+	watchDebounced,
+} from "@vueuse/core";
+import { NDataTable, NPagination, useNotification } from "naive-ui";
 import { storeToRefs } from "pinia";
 import { accessKeys } from "server/schema";
-import { computed, onMounted, ref, watch } from "vue";
+import { computed, ref, watch } from "vue";
+import BAccessDataFilter from "@/components/access-data/BAccessDataFilter.vue";
+import BAccessDataTags from "@/components/access-data/BAccessDataTags.vue";
 import { WS_URL } from "@/consts.ts";
 import { formatColumns } from "@/module/access-data/format.ts";
 import { useStatsStore } from "@/stores/stats.ts";
@@ -26,18 +20,13 @@ const notification = useNotification();
 const statsStore = useStatsStore();
 const { accessLog, total, loading, count, error } = storeToRefs(statsStore);
 
-const fieldOptions = accessKeys.map((key: string) => {
-	return {
-		label: `@${key}:`,
-		value: `@${key}:`,
-	};
-});
-
 const form = ref({
 	field: "@user:",
 	search: null,
 });
 
+const breakpoints = useBreakpoints(breakpointsTailwind);
+const tableMaxHeight = computed(() => (breakpoints.md.value ? 560 : 280));
 const interval = ref(300);
 const pause = ref(false);
 const page = ref();
@@ -77,6 +66,7 @@ watchDebounced(
 watchDebounced(
 	form,
 	async (v) => {
+		page.value = 1;
 		await statsStore.getAccessLogs({
 			page: page.value,
 			search: search.value,
@@ -149,85 +139,28 @@ watchDebounced(
 
 <template>
   <div class="access-data flex flex-col gap-5">
-    <NForm :model="form" class="search">
-      <div class="max-w-xl flex gap-2">
-        <n-select
-            v-model:value="form.field"
-            placeholder="Select"
-            :options="fieldOptions"
-            size="large"
-            class="w-[200px]"
-        />
-        <NInput
-            v-model:value="form.search"
-            placeholder="Search by column"
-            size="large"
-        />
+    <BAccessDataFilter
+        v-model="form"
+    />
 
-        <NTooltip placement="bottom" trigger="hover">
-          <template #trigger>
-            <Icon size="32" class="cursor-pointer">
-              <InformationCircle />
-            </Icon>
-          </template>
-          This input use for <a target="_blank" rel="noopener" href="https://redis.io/docs/latest/develop/ai/search-and-query/query/" class="c-amber underline">redis query</a> like <NTag>@user: fox</NTag>
-          <br>
-          Search by IP: <NTag>192.168.1.1</NTag> (auto-escaped)
-          <br>
-          Single number: <NTag>200</NTag> → <NTag>[200 200]</NTag>
-          <br>
-          Number ranges: <NTag>[200 299]</NTag> or <NTag>[0 20]</NTag>
-        </NTooltip>
-      </div>
-    </NForm>
-    <div class="flex gap-2">
-      <NTag class="text-xl">
-        Total Records: {{ total }}
-      </NTag>
-      <NTag class="text-xl">
-        Count: {{ count }}
-      </NTag>
-      <NButton
-          size="small"
-          :type="`${!pause ? 'error' : 'success'}`"
-          @click="handlePause"
-          class="text-xl"
-      >
-        <template #icon>
-          <Icon>
-            <Pause v-if="!pause" />
-            <Play v-else />
-          </Icon>
-        </template>
+    <BAccessDataTags
+        v-model:interval="interval"
+        v-bind="{ total, count, pause, status }"
+        @handlePause="handlePause"
+    />
 
-        {{ !pause ? 'Pause' : 'Start' }}
-      </NButton>
-
-      <NTag class="text-xl">
-        Interval:
-      </NTag>
-      <NInputNumber
-          v-model:value="interval"
-          size="small"
-          :show-button="false"
-          class="max-w-[50px]"
-      />
-      <NTag class="text-xl">
-        WS: {{ status }}
-      </NTag>
-    </div>
-
-    <n-data-table
+    <NDataTable
         :columns="columns"
         :data="accessLog"
-        :max-height="560"
+        :max-height="tableMaxHeight"
         :loading="loading"
         :row-class-name="rowClassName"
+        :scroll-x="1420"
     />
-    <n-pagination
+    <NPagination
         v-model:page="page"
         :page-count="pageCount"
-        show-quick-jumper
+        simple
         size="large"
     />
   </div>