+++ apps/web/src/components/access-metric/masonry-wall/BCardBytes.vue
@@ -1,0 +1,22 @@
+<script setup lang="ts">
+import BCardMetric from "@/components/BCardMetric.vue";
+import { formatBytes } from "@/utils/string.ts";
+
+const { bytes = 0 } = defineProps<{
+	bytes?: number;
+}>();
+</script>
+
+<template>
+  <BCardMetric>
+    {{ formatBytes(bytes) }}
+
+    <template #name>
+      TOTAL REQUEST SIZE
+    </template>
+  </BCardMetric>
+</template>
+
+<style scoped>
+
+</style>